Refactor code to optimize facade usage
The changes made in this commit improve the utilization of the facade pattern in the test cases. The "PatchAfterGetProb" method now uses the 'this' keyword making it an extension method of the interface 'ICreatureEquipmentMakeInfoFacade', reducing duplication and increasing code readability. This adjustment not only contributes to cleaner code but also to the enhancement of code maintainability.