Add mechanism to reset Crumbling Armor's agent list on gift change
A new method was added to reinitialize the internal list of Crumbling Armor agents whenever any agent changes their gift. This tackles an existing bug, where an agent that started the day with Crumbling Armor's gift but later replaced it, would still die when performing an attachment. Significant changes were also made in ArmorCreaturePatchOnNotice and added IArmorCreatureTestAdapter for tests.