Refactor code and relocate customizingWindowExtensions methods
The CustomizingWindowExtensions methods have been relocated to a common namespace for reusability purposes. In the process, the code was refactored and various functions were cleaned up. The `WorkerModelExtensions.cs` file was moved from `FreeCustomization` to the `Common` directory and the necessary references were adjusted in other files. The method `CustomizingWindowPatchConfirm.PatchBeforeConfirm` was adjusted to utilize the facade pattern.