Fixed image not loading because of test adapter parameters.
The committed changes introduce an ImageParameters object which gathers all image related parameters in one place. Now, instead of supplying each parameter separately, an ImageParameters instance is passed around. This makes the code much cleaner and easier to understand. The changes affect various methods that deal with image creation and updating. This approach also enhances the flexibility of parameter modifications in the future.