Exception logs will now be appended to log all exceptions.
In this update, the behavior of the file writing methods has been updated to support append functionality. Previously, every time you wrote to a file, the existing contents were overwritten. Now, you can choose to append to the existing contents of a file instead. This change affects various parts of the system, including tests and the logging system.