Responses now work correctly in the SDK.
The update refined the CrowdControlResponseStatus enum for clearer meaning and improved readability. Some variable names were changed for more precise semantics. The file reading mechanism in the CrowdControlRequest class was optimised to be more efficient. In addition, an error-catching mechanism was improved to provide more specific error messages. Lastly, the "Example" effect in the CrowdControlDelegates class was replaced with a real "AddEnergy" effect.